+++ core/src/sync/message/get_block_txn_response.rs
@@ -28,7 +28,13 @@ impl Handleable for GetBlockTxnResponse {
     fn handle(self, ctx: &Context) -> Result<(), Error> {
         let _timer = MeterTimer::time_func(BLOCK_TXN_HANDLE_TIMER.as_ref());
 
-        debug!("on_get_blocktxn_response");
+        debug!("on_get_blocktxn_response, hash={:?}", self.block_hash);
+
+        if ctx.manager.is_block_queue_full() {
+            warn!("recover_public_queue is full, discard GetBlockTxnResponse");
+            return Ok(());
+        }
+
         let resp_hash = self.block_hash;
         let req = ctx.match_request(self.request_id)?;
         let delay = req.delay;